feat: implement recipe import via share sheet and clean up codebase
This commit introduces the ability to import recipes by sharing a URL to the app.

- Add intent filter for `ACTION_SEND` to handle shared text (URLs).
- Implement `RecipeImportRepository` and `RecipeImportViewModel` to manage the import process.
- Update `MainActivity` to handle the share intent, display toasts on import status, and add a splash screen.
- Add new launcher icons and splash screen theme.
- Update Gradle version and dependencies.
- Refactor and remove a significant amount of unused code, components, and resources.
- Switch logging to use Timber.

+++ app/src/main/java/com/davidniederweis/mealier/ui/components/recipe/RecipeDetailContent.kt
@@ -40,7 +40,7 @@ fun RecipeDetailContent(
     val completedInstructions = remember { mutableStateSetOf<Int>() }
 
     // Track current servings count for scaling ingredients
-    var currentServings by remember { mutableDoubleStateOf(recipe.recipeServings.toDouble()) }
+    var currentServings by remember { mutableDoubleStateOf(recipe.recipeServings) }
 
     // Calculate servings multiplier from current servings
     val servingsMultiplier = if (recipe.recipeServings > 0) {
